Move library header files to a separated directory
