+++ include/pairg/heuristics.hpp
