+++ tests/test_main.cpp
@@ -7,7 +7,7 @@
 #define CATCH_CONFIG_MAIN
 
 //External includes
-#include "catch/single_include/catch2/catch.hpp"
+#include <catch2/catch.hpp>
 
 TEST_CASE( "1: All test cases reside in other .cpp files (empty)") {
 }
